Initial REST API for nswSiteSetting
