+++ src/nswdesignsystem/plone6/configure.zcml
@@ -14,6 +14,7 @@
   <!--<includeDependencies package="." />-->
   <include file="dependencies.zcml" />
   <include package=".browser" />
+  <include package=".restapi" />
 
   <include file="permissions.zcml" />
