+++ src/nswdesignsystem/plone6/restapi/serializer/configure.zcml
@@ -1,0 +1,7 @@
+<configure xmlns="http://namespaces.zope.org/zope"
+  xmlns:plone="http://namespaces.plone.org/plone"
+  xmlns:zcml="http://namespaces.zope.org/zcml">
+
+  <adapter factory=".nsw_site_settings.INSWDesignSystemSettingsSerializeToJson" />
+
+</configure>
